+++ src/lib/converter/comments/blockLexer.ts
@@ -201,7 +201,10 @@ function* lexBlockComment2(
                     }
                 }
 
-                if (lookahead !== pos + 1) {
+                if (
+                    lookahead !== pos + 1 &&
+                    (lookahead === end || /\s/.test(file[lookahead]))
+                ) {
                     braceStartsType = true;
                     yield makeToken(TokenSyntaxKind.Tag, lookahead - pos);
                     break;